# Pack style into uint32