--- a/go.mod
+++ b/go.mod
@@ -1,6 +1,6 @@
 module github.com/buildkite/terminal-to-html/v3
 
-go 1.20
+go 1.21
 
 require (
 	github.com/google/go-cmp v0.5.9