--- a/style.go
+++ b/style.go
@@ -2,20 +2,45 @@ package terminal
 
 import "strconv"
 
-var emptyStyle = style{}
+type style uint32
 
-type style struct {
-	fgColor   uint8
-	bgColor   uint8
-	fgColorX  bool
-	bgColorX  bool
-	bold      bool
-	faint     bool
-	italic    bool
-	underline bool
-	strike    bool
-	blink     bool
-}
+// style encoding:
+// 0...  ...7  8... ...15  16...23  24....31
+// [fg color]  [bg color]  [flags]  [unused]
+// flags = bold, faint, etc
+
+const (
+	sbFGColorX = 1 << (16 + iota)
+	sbBGColorX
+	sbBold
+	sbFaint
+	sbItalic
+	sbUnderline
+	sbStrike
+	sbBlink
+)
+
+func (s style) fgColor() uint8  { return uint8(s & 0xff) }
+func (s style) bgColor() uint8  { return uint8((s & 0xff_00) >> 8) }
+func (s style) fgColorX() bool  { return s&sbFGColorX != 0 }
+func (s style) bgColorX() bool  { return s&sbBGColorX != 0 }
+func (s style) bold() bool      { return s&sbBold != 0 }
+func (s style) faint() bool     { return s&sbFaint != 0 }
+func (s style) italic() bool    { return s&sbItalic != 0 }
+func (s style) underline() bool { return s&sbUnderline != 0 }
+func (s style) strike() bool    { return s&sbStrike != 0 }
+func (s style) blink() bool     { return s&sbBlink != 0 }
+
+func (s *style) setFGColor(v uint8)  { *s = (*s &^ 0xff) | style(v) }
+func (s *style) setBGColor(v uint8)  { *s = (*s &^ 0xff_00) | (style(v) << 8) }
+func (s *style) setFGColorX(v bool)  { *s = (*s &^ sbFGColorX) | booln(v, sbFGColorX) }
+func (s *style) setBGColorX(v bool)  { *s = (*s &^ sbBGColorX) | booln(v, sbBGColorX) }
+func (s *style) setBold(v bool)      { *s = (*s &^ sbBold) | booln(v, sbBold) }
+func (s *style) setFaint(v bool)     { *s = (*s &^ sbFaint) | booln(v, sbFaint) }
+func (s *style) setItalic(v bool)    { *s = (*s &^ sbItalic) | booln(v, sbItalic) }
+func (s *style) setUnderline(v bool) { *s = (*s &^ sbUnderline) | booln(v, sbUnderline) }
+func (s *style) setStrike(v bool)    { *s = (*s &^ sbStrike) | booln(v, sbStrike) }
+func (s *style) setBlink(v bool)     { *s = (*s &^ sbBlink) | booln(v, sbBlink) }
 
 const (
 	COLOR_NORMAL        = iota
@@ -25,75 +50,61 @@ const (
 	COLOR_GOT_48        = iota
 )
 
-// True if both styles are equal (or are the same object)
-func (s *style) isEqual(o *style) bool {
-	return s == o || *s == *o
-}
-
 // CSS classes that make up the style
-func (s *style) asClasses() []string {
+func (s style) asClasses() []string {
 	var styles []string
 
-	if s.fgColor > 0 && s.fgColor < 38 && !s.fgColorX {
-		styles = append(styles, "term-fg"+strconv.Itoa(int(s.fgColor)))
+	if s.fgColor() > 0 && s.fgColor() < 38 && !s.fgColorX() {
+		styles = append(styles, "term-fg"+strconv.Itoa(int(s.fgColor())))
 	}
-	if s.fgColor > 38 && !s.fgColorX {
-		styles = append(styles, "term-fgi"+strconv.Itoa(int(s.fgColor)))
+	if s.fgColor() > 38 && !s.fgColorX() {
+		styles = append(styles, "term-fgi"+strconv.Itoa(int(s.fgColor())))
 
 	}
-	if s.fgColorX {
-		styles = append(styles, "term-fgx"+strconv.Itoa(int(s.fgColor)))
+	if s.fgColorX() {
+		styles = append(styles, "term-fgx"+strconv.Itoa(int(s.fgColor())))
 
 	}
 
-	if s.bgColor > 0 && s.bgColor < 48 && !s.bgColorX {
-		styles = append(styles, "term-bg"+strconv.Itoa(int(s.bgColor)))
+	if s.bgColor() > 0 && s.bgColor() < 48 && !s.bgColorX() {
+		styles = append(styles, "term-bg"+strconv.Itoa(int(s.bgColor())))
 	}
-	if s.bgColor > 48 && !s.bgColorX {
-		styles = append(styles, "term-bgi"+strconv.Itoa(int(s.bgColor)))
+	if s.bgColor() > 48 && !s.bgColorX() {
+		styles = append(styles, "term-bgi"+strconv.Itoa(int(s.bgColor())))
 	}
-	if s.bgColorX {
-		styles = append(styles, "term-bgx"+strconv.Itoa(int(s.bgColor)))
+	if s.bgColorX() {
+		styles = append(styles, "term-bgx"+strconv.Itoa(int(s.bgColor())))
 	}
 
-	if s.bold {
+	if s.bold() {
 		styles = append(styles, "term-fg1")
 	}
-	if s.faint {
+	if s.faint() {
 		styles = append(styles, "term-fg2")
 	}
-	if s.italic {
+	if s.italic() {
 		styles = append(styles, "term-fg3")
 	}
-	if s.underline {
+	if s.underline() {
 		styles = append(styles, "term-fg4")
 	}
-	if s.blink {
+	if s.blink() {
 		styles = append(styles, "term-fg5")
 	}
-	if s.strike {
+	if s.strike() {
 		styles = append(styles, "term-fg9")
 	}
 
 	return styles
 }
 
-// True if style is empty
-func (s *style) isEmpty() bool {
-	return *s == style{}
-}
-
-// Add colours to an existing style, potentially returning
-// a new style.
-func (s *style) color(colors []string) *style {
+// Add colours to an existing style, returning a new style.
+func (s style) color(colors []string) style {
 	if len(colors) == 1 && (colors[0] == "0" || colors[0] == "") {
-		// Shortcut for full style reset
-		return &emptyStyle
+		return 0
 	}
 
-	newStyle := style(*s)
-	s = &newStyle
-	color_mode := COLOR_NORMAL
+	colorMode := COLOR_NORMAL
 
 	for _, ccs := range colors {
 		// If multiple colors are defined, i.e. \e[30;42m\e then loop through each
@@ -104,80 +115,87 @@ func (s *style) color(colors []string) *style {
 		}
 
 		// State machine for XTerm colors, eg 38;5;150
-		switch color_mode {
+		switch colorMode {
 		case COLOR_GOT_38_NEED_5:
 			if cc == 5 {
-				color_mode = COLOR_GOT_38
+				colorMode = COLOR_GOT_38
 			} else {
-				color_mode = COLOR_NORMAL
+				colorMode = COLOR_NORMAL
 			}
 			continue
 		case COLOR_GOT_48_NEED_5:
 			if cc == 5 {
-				color_mode = COLOR_GOT_48
+				colorMode = COLOR_GOT_48
 			} else {
-				color_mode = COLOR_NORMAL
+				colorMode = COLOR_NORMAL
 			}
 			continue
 		case COLOR_GOT_38:
-			s.fgColor = uint8(cc)
-			s.fgColorX = true
-			color_mode = COLOR_NORMAL
+			s.setFGColor(uint8(cc))
+			s.setFGColorX(true)
+			colorMode = COLOR_NORMAL
 			continue
 		case COLOR_GOT_48:
-			s.bgColor = uint8(cc)
-			s.bgColorX = true
-			color_mode = COLOR_NORMAL
+			s.setBGColor(uint8(cc))
+			s.setBGColorX(true)
+			colorMode = COLOR_NORMAL
 			continue
 		}
 
 		switch cc {
 		case 0:
-			// Reset all styles - don't use &emptyStyle here as we could end up adding colours
-			// in this same action.
-			s = &style{}
+			// Reset all styles
+			s = 0
 		case 1:
-			s.bold = true
-			s.faint = false
+			s.setBold(true)
+			s.setFaint(false)
 		case 2:
-			s.faint = true
-			s.bold = false
+			s.setFaint(true)
+			s.setBold(false)
 		case 3:
-			s.italic = true
+			s.setItalic(true)
 		case 4:
-			s.underline = true
+			s.setUnderline(true)
 		case 5, 6:
-			s.blink = true
+			s.setBlink(true)
 		case 9:
-			s.strike = true
+			s.setStrike(true)
 		case 21, 22:
-			s.bold = false
-			s.faint = false
+			s.setBold(false)
+			s.setFaint(false)
 		case 23:
-			s.italic = false
+			s.setItalic(false)
 		case 24:
-			s.underline = false
+			s.setUnderline(false)
 		case 25:
-			s.blink = false
+			s.setBlink(false)
 		case 29:
-			s.strike = false
+			s.setStrike(false)
 		case 38:
-			color_mode = COLOR_GOT_38_NEED_5
+			colorMode = COLOR_GOT_38_NEED_5
 		case 39:
-			s.fgColor = 0
-			s.fgColorX = false
+			s.setFGColor(0)
+			s.setFGColorX(false)
 		case 48:
-			color_mode = COLOR_GOT_48_NEED_5
+			colorMode = COLOR_GOT_48_NEED_5
 		case 49:
-			s.bgColor = 0
-			s.bgColorX = false
+			s.setBGColor(0)
+			s.setBGColorX(false)
 		case 30, 31, 32, 33, 34, 35, 36, 37, 90, 91, 92, 93, 94, 95, 96, 97:
-			s.fgColor = uint8(cc)
-			s.fgColorX = false
+			s.setFGColor(uint8(cc))
+			s.setFGColorX(false)
 		case 40, 41, 42, 43, 44, 45, 46, 47, 100, 101, 102, 103, 104, 105, 106, 107:
-			s.bgColor = uint8(cc)
-			s.bgColorX = false
+			s.setBGColor(uint8(cc))
+			s.setBGColorX(false)
 		}
 	}
 	return s
+}
+
+// false, true => 0, t
+func booln(b bool, t style) style {
+	if b {
+		return t
+	}
+	return 0
 }