The old node consists of the rune (`blob`), which is an int32, plus two pointers. On a 32-bit platform these align nicely. On 64-bit, the compiler typically wastes 4 bytes on padding for alignment.

Storing style as a packed int, and replacing *style with style, has a few benefits:
- blob + style is 8 bytes, and the remaining pointer takes it to 16 bytes
- comparison of styles is now a single integer comparison
- emptyStyle is now just 0

Also it's faster and uses less memory:

```plain
goos: linux
goarch: amd64
pkg: github.com/buildkite/terminal-to-html/v3
cpu: Intel(R) Xeon(R) CPU E5-1650 v2 @ 3.50GHz
                      │ baseline.txt │             packed.txt              │
                      │    sec/op    │   sec/op     vs base                │
RendererControl-12       3.631µ ± 1%   2.898µ ± 1%  -20.19% (p=0.000 n=30)
RendererCurl-12          50.02µ ± 0%   46.34µ ± 0%   -7.34% (p=0.000 n=30)
RendererHomer-12        104.55µ ± 0%   90.86µ ± 0%  -13.09% (p=0.000 n=30)
RendererDockerPull-12    181.2µ ± 0%   167.9µ ± 0%   -7.37% (p=0.000 n=30)
RendererPikachu-12       5.030m ± 0%   4.766m ± 0%   -5.26% (p=0.000 n=30)
RendererNpm-12           80.98m ± 2%   81.00m ± 1%        ~ (p=0.697 n=30)
geomean                  334.5µ        304.1µ        -9.10%

                      │ baseline.txt  │              packed.txt              │
                      │     B/op      │     B/op      vs base                │
RendererControl-12       2.227Ki ± 0%   1.477Ki ± 0%  -33.68% (p=0.000 n=30)
RendererCurl-12         10.766Ki ± 0%   7.750Ki ± 0%  -28.01% (p=0.000 n=30)
RendererHomer-12         46.91Ki ± 0%   33.39Ki ± 0%  -28.81% (p=0.000 n=30)
RendererDockerPull-12    48.41Ki ± 0%   33.66Ki ± 0%  -30.47% (p=0.000 n=30)
RendererPikachu-12       767.6Ki ± 0%   652.4Ki ± 0%  -15.01% (p=0.000 n=30)
RendererNpm-12           24.26Mi ± 0%   18.44Mi ± 0%  -23.99% (p=0.000 n=30)
geomean                  100.6Ki        73.56Ki       -26.89%

                      │ baseline.txt │              packed.txt               │
                      │  allocs/op   │  allocs/op   vs base                  │
RendererControl-12        9.000 ± 0%    9.000 ± 0%        ~ (p=1.000 n=30) ¹
RendererCurl-12           36.00 ± 0%    35.00 ± 0%   -2.78% (p=0.000 n=30)
RendererHomer-12          134.0 ± 0%    133.0 ± 0%   -0.75% (p=0.000 n=30)
RendererDockerPull-12     193.0 ± 0%    193.0 ± 0%        ~ (p=1.000 n=30) ¹
RendererPikachu-12       16.16k ± 0%   14.35k ± 0%  -11.23% (p=0.000 n=30)
RendererNpm-12           183.2k ± 0%   158.9k ± 0%  -13.27% (p=0.000 n=30)
geomean                   540.1         514.0        -4.83%
¹ all samples are equal
```

```plain
goos: darwin
goarch: arm64
pkg: github.com/buildkite/terminal-to-html/v3
                      │ baseline-darwin-arm64.txt │       packed-darwin-arm64.txt       │
                      │          sec/op           │   sec/op     vs base                │
RendererControl-10                    620.6n ± 5%   556.3n ± 1%  -10.35% (p=0.000 n=30)
RendererCurl-10                       8.306µ ± 0%   7.644µ ± 0%   -7.98% (p=0.000 n=30)
RendererHomer-10                      19.65µ ± 0%   16.55µ ± 1%  -15.81% (p=0.000 n=30)
RendererDockerPull-10                 32.65µ ± 0%   28.96µ ± 2%  -11.30% (p=0.000 n=30)
RendererPikachu-10                    739.5µ ± 0%   682.2µ ± 1%   -7.75% (p=0.000 n=30)
RendererNpm-10                        15.24m ± 0%   13.43m ± 0%  -11.85% (p=0.000 n=30)
geomean                               57.79µ        51.51µ       -10.88%

                      │ baseline-darwin-arm64.txt │       packed-darwin-arm64.txt        │
                      │           B/op            │     B/op      vs base                │
RendererControl-10                   2.227Ki ± 0%   1.477Ki ± 0%  -33.68% (p=0.000 n=30)
RendererCurl-10                     10.766Ki ± 0%   7.750Ki ± 0%  -28.01% (p=0.000 n=30)
RendererHomer-10                     46.91Ki ± 0%   33.39Ki ± 0%  -28.81% (p=0.000 n=30)
RendererDockerPull-10                48.41Ki ± 0%   33.66Ki ± 0%  -30.47% (p=0.000 n=30)
RendererPikachu-10                   767.6Ki ± 0%   652.4Ki ± 0%  -15.01% (p=0.000 n=30)
RendererNpm-10                       24.26Mi ± 0%   18.44Mi ± 0%  -23.99% (p=0.000 n=30)
geomean                              100.6Ki        73.56Ki       -26.89%

                      │ baseline-darwin-arm64.txt │        packed-darwin-arm64.txt        │
                      │         allocs/op         │  allocs/op   vs base                  │
RendererControl-10                     9.000 ± 0%    9.000 ± 0%        ~ (p=1.000 n=30) ¹
RendererCurl-10                        36.00 ± 0%    35.00 ± 0%   -2.78% (p=0.000 n=30)
RendererHomer-10                       134.0 ± 0%    133.0 ± 0%   -0.75% (p=0.000 n=30)
RendererDockerPull-10                  193.0 ± 0%    193.0 ± 0%        ~ (p=1.000 n=30) ¹
RendererPikachu-10                    16.16k ± 0%   14.35k ± 0%  -11.23% (p=0.000 n=30)
RendererNpm-10                        183.2k ± 0%   158.9k ± 0%  -13.27% (p=0.000 n=30)
geomean                                540.1         514.0        -4.83%
¹ all samples are equal
```

```plain
goos: darwin
goarch: amd64
pkg: github.com/buildkite/terminal-to-html/v3
cpu: Intel(R) Xeon(R) W-2191B CPU @ 2.30GHz
                      │ baseline.txt │             packed.txt              │
                      │    sec/op    │   sec/op     vs base                │
RendererControl-36       1.606µ ± 0%   1.172µ ± 1%  -27.00% (p=0.000 n=30)
RendererCurl-36          14.29µ ± 0%   12.52µ ± 0%  -12.36% (p=0.000 n=30)
RendererHomer-36         40.22µ ± 0%   32.54µ ± 0%  -19.09% (p=0.000 n=30)
RendererDockerPull-36    54.76µ ± 0%   47.35µ ± 0%  -13.52% (p=0.000 n=30)
RendererPikachu-36       1.312m ± 0%   1.212m ± 0%   -7.64% (p=0.000 n=30)
RendererNpm-36           23.35m ± 1%   21.53m ± 0%   -7.76% (p=0.000 n=30)
geomean                  107.5µ        91.58µ       -14.84%

                      │ baseline.txt  │              packed.txt              │
                      │     B/op      │     B/op      vs base                │
RendererControl-36       2.227Ki ± 0%   1.477Ki ± 0%  -33.68% (p=0.000 n=30)
RendererCurl-36         10.766Ki ± 0%   7.750Ki ± 0%  -28.01% (p=0.000 n=30)
RendererHomer-36         46.91Ki ± 0%   33.39Ki ± 0%  -28.81% (p=0.000 n=30)
RendererDockerPull-36    48.41Ki ± 0%   33.66Ki ± 0%  -30.47% (p=0.000 n=30)
RendererPikachu-36       767.6Ki ± 0%   652.4Ki ± 0%  -15.01% (p=0.000 n=30)
RendererNpm-36           24.26Mi ± 0%   18.44Mi ± 0%  -23.99% (p=0.000 n=30)
geomean                  100.6Ki        73.56Ki       -26.89%

                      │ baseline.txt │              packed.txt               │
                      │  allocs/op   │  allocs/op   vs base                  │
RendererControl-36        9.000 ± 0%    9.000 ± 0%        ~ (p=1.000 n=30) ¹
RendererCurl-36           36.00 ± 0%    35.00 ± 0%   -2.78% (p=0.000 n=30)
RendererHomer-36          134.0 ± 0%    133.0 ± 0%   -0.75% (p=0.000 n=30)
RendererDockerPull-36     193.0 ± 0%    193.0 ± 0%        ~ (p=1.000 n=30) ¹
RendererPikachu-36       16.16k ± 0%   14.35k ± 0%  -11.23% (p=0.000 n=30)
RendererNpm-36           183.2k ± 0%   158.9k ± 0%  -13.27% (p=0.000 n=30)
geomean                   540.1         514.0        -4.83%
¹ all samples are equal
```